Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
faiss-cpu
sentence-transformers
transformers
peft
accelerate
bitsandbytes
torch
requests
